Add gelu.py selu.py dropout.py dtype supported

+++ gelu.py
@@ -3,7 +3,6 @@ import triton
 import triton.language as tl
 from .libentry import libentry
 
-
 @libentry()
 @triton.autotune(configs=[
     triton.Config({"N_BLOCK_SIZE": 256}, num_warps=2, num_stages=4),
@@ -43,7 +42,7 @@ def gelu_kernel(
     output = 0.5 * input * (
         1 + tl.math.tanh(input * 0.79788456 *
                          (1 + 0.044715 * tl.math.pow(input.to(tl.float32), 2))))
-    tl.store(Y_ptrs, output.to(tl.float16))
+    tl.store(Y_ptrs, output.to(input.dtype))
     
 
 def gelu(A, *, out=None):
